# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GAC授权中心",
        "title": "GAC授权中心",
        "termsOfService": "https://paymedia.yuque.com/os/hb9uhc",
        "contact": {
            "name": "org.pm.gac.service",
            "url": "https://www.apifox.cn/web/main"
        },
        "license": {
            "name": "GIN.swagger",
            "url": "https://github.com/swaggo/swag/blob/master/README_zh-CN.md#%E5%A6%82%E4%BD%95%E4%B8%8Egin%E9%9B%86%E6%88%90"
        },
        "version": "2.0"
    },
    "host": "api.t.dev.pay.fun",
    "basePath": "/gac/v1",
    "paths": {
...
        "/admin/menu/create": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/菜单管理"
                ],
                "summary": "增加菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.AddMenuReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/delete": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/菜单管理"
                ],
                "summary": "菜单删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdateMenuStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/disable": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/菜单管理"
                ],
                "summary": "菜单禁用/启用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdateMenuStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/page": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/菜单管理"
                ],
                "summary": "菜单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.MenuReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/adminApi.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/assembler.MenuAssembler"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/tree": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/菜单管理"
                ],
                "summary": "菜单树形结构",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/adminApi.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/lib.Tree"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/update": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/菜单管理"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdateMenuReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "adminApi.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "assembler.MenuAssembler": {
            "type": "object",
            "properties": {
                "api_route": {
                    "description": "接口路由",
                    "type": "string"
                },
                "app_code": {
                    "description": "组织Code",
                    "type": "string"
                },
                "app_name": {
                    "description": "组织名称",
                    "type": "string"
                },
                "app_pk": {
                    "description": "应用",
                    "type": "integer"
                },
                "code": {
                    "description": "权限code",
                    "type": "string"
                },
                "created_ip": {
                    "description": "创建者Ip",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "created_user": {
                    "description": "创建人",
                    "type": "string"
                },
                "i18n_code": {
                    "description": "多语言CODE",
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "joins": {
                    "description": "id路径",
                    "type": "string"
                },
                "keep_alive": {
                    "description": "路由缓存： 1：是 2：否",
                    "type": "integer"
                },
                "method": {
                    "description": "请求方法:get,post,put,patch,options",
                    "type": "string"
                },
                "name": {
                    "description": "权限名称",
                    "type": "string"
                },
                "org_name": {
                    "description": "应用名称",
                    "type": "string"
                },
                "org_pk": {
                    "description": "组织",
                    "type": "integer"
                },
                "parent_name": {
                    "description": "上级名称",
                    "type": "string"
                },
                "parent_pk": {
                    "description": "上级id",
                    "type": "integer"
                },
                "pk": {
                    "type": "integer"
                },
                "redirect": {
                    "description": "前台路由重定向",
                    "type": "string"
                },
                "route": {
                    "description": "前台路由",
                    "type": "string"
                },
                "show": {
                    "description": "前端显示状态1：显示 2：隐藏",
                    "type": "integer"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1:启用 2：停用  3:  删除",
                    "type": "integer"
                },
                "type": {
                    "description": "菜单类型： 1：菜单 2：路由 3：按钮",
                    "type": "integer"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                },
                "version": {
                    "description": "版本",
                    "type": "string"
                },
                "view_path": {
                    "description": "前端模板路径",
                    "type": "string"
                }
            }
        },
...
        "lib.Tree": {
            "type": "object",
            "properties": {
                "checked": {
                    "description": "选中",
                    "type": "boolean"
                },
                "children": {
                    "description": "子节点",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Tree"
                    }
                },
                "data": {
                    "description": "自定义对象",
                    "type": "object"
                },
                "leaf": {
                    "description": "叶子节点",
                    "type": "boolean"
                },
                "parent_pk": {
                    "description": "节点父级Id",
                    "type": "integer"
                },
                "partial_selected": {
                    "description": "部分选中",
                    "type": "boolean"
                },
                "pk": {
                    "description": "节点Id",
                    "type": "integer"
                },
                "title": {
                    "description": "节点名字",
                    "type": "string"
                }
            }
        },
...
        "req.AddMenuReq": {
            "type": "object",
            "required": [
                "name",
                "show",
                "status",
                "type"
            ],
            "properties": {
                "api_route": {
                    "description": "接口路由",
                    "type": "string"
                },
                "app_pk": {
                    "description": "应用",
                    "type": "integer"
                },
                "code": {
                    "description": "权限code",
                    "type": "string"
                },
                "created_ip": {
                    "description": "创建者Ip",
                    "type": "string"
                },
                "created_user": {
                    "description": "创建人",
                    "type": "string"
                },
                "i18n_code": {
                    "description": "多语言CODE",
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "keep_alive": {
                    "description": "路由缓存： 1：是 2：否",
                    "type": "integer"
                },
                "method": {
                    "description": "请求方法:get,post,put,patch,options",
                    "type": "string"
                },
                "name": {
                    "description": "权限名称",
                    "type": "string"
                },
                "org_pk": {
                    "description": "组织",
                    "type": "integer"
                },
                "parent_pk": {
                    "description": "上级id",
                    "type": "integer"
                },
                "redirect": {
                    "description": "前台路由重定向",
                    "type": "string"
                },
                "route": {
                    "description": "前台路由",
                    "type": "string"
                },
                "show": {
                    "description": "前端显示状态 1：显示 2：隐藏",
                    "type": "integer"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1:启用 2：停用  3:  删除",
                    "type": "integer"
                },
                "type": {
                    "description": "菜单类型： 1：菜单 2：路由 3：按钮",
                    "type": "integer"
                },
                "view_path": {
                    "description": "前端模板路径",
                    "type": "string"
                }
            }
        },
...
        "req.MenuReq": {
            "type": "object",
            "properties": {
                "app_pk": {
                    "description": "应用ID",
                    "type": "integer"
                },
                "current": {
                    "description": "当前页",
                    "type": "integer"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "org_pk": {
                    "description": "组织ID",
                    "type": "integer"
                },
                "page_size": {
                    "description": "页大小",
                    "type": "integer"
                },
                "pagination": {
                    "description": "是否使用分页查询",
                    "type": "boolean"
                },
                "parent_pk": {
                    "description": "父级Pk",
                    "type": "string"
                },
                "pk": {
                    "description": "pk",
                    "type": "integer"
                },
                "pks": {
                    "description": "pks",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "role_pk": {
                    "description": "角色ID",
                    "type": "integer"
                },
                "role_type": {
                    "description": "角色类型 1:管理员 2：员工 3： 服务商",
                    "type": "integer"
                },
                "statues": {
                    "description": "状态：1:启用 2：停用  3:  删除 status 数组形式",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "description": "状态：1:启用 2：停用  3:  删除",
                    "type": "integer"
                },
                "type": {
                    "description": "type",
                    "type": "integer"
                },
                "types": {
                    "description": "types  type数组形式",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "req.UpdateMenuReq": {
            "type": "object",
            "required": [
                "name",
                "show",
                "status",
                "type"
            ],
            "properties": {
                "api_route": {
                    "description": "接口路由",
                    "type": "string"
                },
                "app_pk": {
                    "description": "应用",
                    "type": "integer"
                },
                "code": {
                    "description": "权限code",
                    "type": "string"
                },
                "created_ip": {
                    "description": "创建者Ip",
                    "type": "string"
                },
                "created_user": {
                    "description": "创建人",
                    "type": "string"
                },
                "i18n_code": {
                    "description": "多语言CODE",
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "keep_alive": {
                    "description": "路由缓存： 1：是 2：否",
                    "type": "integer"
                },
                "method": {
                    "description": "请求方法:get,post,put,patch,options",
                    "type": "string"
                },
                "name": {
                    "description": "权限名称",
                    "type": "string"
                },
                "org_pk": {
                    "description": "组织",
                    "type": "integer"
                },
                "parent_pk": {
                    "description": "上级id",
                    "type": "integer"
                },
                "pk": {
                    "description": "Pk",
                    "type": "integer"
                },
                "redirect": {
                    "description": "前台路由重定向",
                    "type": "string"
                },
                "route": {
                    "description": "前台路由",
                    "type": "string"
                },
                "show": {
                    "description": "前端显示状态 1：显示 2：隐藏",
                    "type": "integer"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1:启用 2：停用  3:  删除",
                    "type": "integer"
                },
                "type": {
                    "description": "菜单类型： 1：菜单 2：路由 3：按钮",
                    "type": "integer"
                },
                "view_path": {
                    "description": "前端模板路径",
                    "type": "string"
                }
            }
        },
        "req.UpdateMenuStatusReq": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "created_ip": {
                    "description": "创建者Ip",
                    "type": "string"
                },
                "created_user": {
                    "description": "创建人",
                    "type": "string"
                },
                "pk": {
                    "description": "Pk",
                    "type": "integer"
                },
                "pks": {
                    "description": "pks",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "description": "状态：1:启用 2：停用  3:  删除",
                    "type": "integer"
                }
            }
        },
...
    }
}